+++ models/shopping_model.py
@@ -1,6 +1,7 @@
 from __future__ import annotations
 
 from datetime import date
+from decimal import Decimal
 
 from pydantic import BaseModel, Field
 
@@ -8,7 +9,7 @@ from pydantic import BaseModel, Field
 class ShoppingItem(BaseModel):
     id: int | None = None
     name: str = Field(min_length=1)
-    price: float = Field(gt=0)
+    price: Decimal = Field(gt=0)
     category: str
     purchased: bool = False
     purchase_date: date | None = None
